#3 - build and test fixes
1. Fixed plugin imports (StackOverflow rides again!)
2. Updated pydantic deprecated stuff.
3. Moved RapidSwarm out of __main__.py.
4. Moved tests directory to correct level.
5. Added simple Makefile b/c one forgets the Poetry commands.
