+++ src/plugins/reporters/reporter_csv_plugin.py
@@ -2,7 +2,7 @@ import csv
 from datetime import datetime
 from typing import List, Union
 
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 
 from rapidswarm.models.reporters import BaseReporter
 
@@ -10,10 +10,9 @@ __module_name__ = "reporter_csv_plugin"
 
 
 class CSVReporter(BaseReporter):
-    output_file: str = "report.csv"
+    model_config = ConfigDict(arbitrary_types_allowed=True)
 
-    class Config:
-        arbitrary_types_allowed = True
+    output_file: str = "report.csv"
 
     def report(self, data: Union[BaseModel, List[BaseModel]]):
         if issubclass(type(data), BaseModel):
@@ -27,7 +26,7 @@ class CSVReporter(BaseReporter):
                 "Unsupported data format. Expected a Pydantic model or a list of Pydantic models."
             )
 
-        fieldnames = ["timestamp"] + list(data_list[0].__fields__.keys())
+        fieldnames = ["timestamp"] + list(data_list[0].model_fields.keys())
 
         try:
             with open(self.output_file, "w", newline="") as file:
@@ -36,7 +35,7 @@ class CSVReporter(BaseReporter):
 
                 for item in data_list:
                     row = {"timestamp": datetime.now().isoformat()}
-                    row.update(item.dict())
+                    row.update(item.model_dump())
                     writer.writerow(row)
         except IOError as e:
             raise IOError(f"Error writing CSV report: {e}")
