+++ tests/test_network_interface.py
