+++ tests/test_network_interface_type.py
