+++ tests/test_scanners.py
